refactor: CircularLabelButton -> IconLabelButton으로 이름 변경
